Add feature flags
