+++ crates/cashu/Cargo.toml
@@ -6,10 +6,14 @@ description = "Cashu shared types and crypto utilities, used as the foundation f
 rust-version = "1.63.0"                                                                                      # MSRV
 
 [features]
+default = ["mint", "wallet"]
 swagger = ["dep:utoipa"]
+mint = ["dep:uuid"]
+wallet = []
 bench = []
 
 [dependencies]
+uuid = { version = "1", features = ["v4", "serde"], optional = true }
 bitcoin = { version = "0.32.2", features = [
     "base64",
     "serde",
@@ -24,7 +28,6 @@ lightning-invoice = { version = "0.32.0", features = ["serde", "std"] }
 thiserror = "2"
 tracing = "0.1"
 url = "2.3"
-uuid = { version = "1", features = ["v4", "serde"] }
 utoipa = { version = "4", optional = true }
 serde_json = "1"
 serde_with = "3"
