+++ crates/cashu/src/lib.rs
@@ -2,11 +2,14 @@
 //!
 pub mod amount;
 pub mod dhke;
+#[cfg(feature = "mint")]
 pub mod mint;
 pub mod mint_url;
 pub mod nuts;
 pub mod secret;
 pub mod util;
+#[cfg(feature = "wallet")]
+pub mod wallet;
 
 pub use lightning_invoice::{self, Bolt11Invoice};
