+++ crates/cdk-common/Cargo.toml
@@ -6,8 +6,11 @@ description = "CDK common types and traits"
 rust-version = "1.63.0"                     # MSRV
 
 [features]
+default = ["mint", "wallet"]
 swagger = ["dep:utoipa", "cashu/swagger"]
 bench = []
+wallet = ["cashu/wallet"]
+mint = ["cashu/mint", "dep:uuid"]
 
 [dependencies]
 async-trait = "0.1"
@@ -17,7 +20,7 @@ bitcoin = { version = "0.32.2", features = [
     "rand",
     "rand-std",
 ] }
-cashu = { path = "../cashu", version = "0.6.0" }
+cashu = { path = "../cashu", default-features = false, version = "0.6.0" }
 cbor-diag = "0.1.12"
 ciborium = { version = "0.2.2", default-features = false, features = ["std"] }
 once_cell = "1.20.2"
@@ -26,7 +29,7 @@ lightning-invoice = { version = "0.32.0", features = ["serde", "std"] }
 thiserror = "2"
 tracing = "0.1"
 url = "2.3"
-uuid = { version = "1", features = ["v4", "serde"] }
+uuid = { version = "1", features = ["v4", "serde"], optional = true }
 utoipa = { version = "4", optional = true }
 futures = "0.3.31"
 anyhow = "1.0"
